"add optimizer naive link option"
