+++ go/pserver/cclient/test/CMakeLists.txt
@@ -1,7 +1,5 @@
 cmake_minimum_required(VERSION 3.0)
 
-add_executable(main main.c)
-add_dependencies(main paddle_pserver_cclient)
 add_executable(test_cclient test_cclient.c)
 add_dependencies(test_cclient paddle_pserver_cclient)
 
@@ -13,10 +11,8 @@ endif()
 
 if(PROJ_ROOT)
   include_directories(${CMAKE_CURRENT_BINARY_DIR}/..)
-  target_link_libraries(main ${CMAKE_CURRENT_BINARY_DIR}/../libpaddle_pserver_cclient.a pthread)
   target_link_libraries(test_cclient ${CMAKE_CURRENT_BINARY_DIR}/../libpaddle_pserver_cclient.a pthread)
 else(PROJ_ROOT)
   include_directories(${CMAKE_BINARY_DIR})
-  target_link_libraries(main ${CMAKE_BINARY_DIR}/libpaddle_pserver_cclient.a pthread)
   target_link_libraries(test_cclient ${CMAKE_BINARY_DIR}/libpaddle_pserver_cclient.a pthread)
 endif(PROJ_ROOT)
