+++ go/pserver/optimizer.go
@@ -1,7 +1,10 @@
 package pserver
 
 /*
+// TODO(zhihong): move compile flags to cmake go_library
+#cgo pkg-config: protobuf
 #cgo CFLAGS: -I ../../
+#cgo LDFLAGS: ../../build/paddle/optimizer/libpaddle_optimizer.a ../../build/proto/libpaddle_proto.a ../../third_party/install/glog/lib/libglog.a ../../third_party/install/gtest/lib/libgtest.a ../../third_party/install/gflags/lib/libgflags.a ../../third_party/install/openblas/lib/libopenblas.a -I/usr/local/lib/ -lprotobuf
 #include "paddle/optimizer/optimizer.h"
 */
 import "C"
